refactor(analyzer): extract callee-ignore tables/predicates into call-graph-builtins.ts (slice 3e)
Next safe slice of modularize-call-graph-builder. Moves the callee-filtering sub-part
of the grab-bag CONSTANTS section (the *_IGNORED per-language tables,
IGNORED_BY_LANGUAGE / ALL_IGNORED_CALLEES, isIgnoredCallee, SELF_CALL_RECEIVERS,
isSelfReceiver) out of call-graph.ts into call-graph-builtins.ts — pure data + string
predicates, zero deps, zero state. The safest remaining slice (no module state, no
re-export juggling).

File-internal: only the two predicates (isIgnoredCallee, isSelfReceiver) are imported
back; the *_IGNORED tables stay private. HUB_THRESHOLD and the style-tally helper stay
put (different concerns). call-graph.ts: 4,846 → 4,741 lines.

Behavior-preserving, verified four ways:
- Snapshot oracle FIRST extended with ignored-builtin calls (print/len/JSON/Math) so
  isIgnoredCallee's drop-or-keep outcome is captured (print/len/JSON dropped as ignored;
  Math.max kept), then captured before/after: IDENTICAL — SHA-256 5fbe0719…
- Public export surface byte-identical (0 added, 0 removed).
- tsc compiles all importers clean; build/lint/typecheck green.
- Full suite green: 279 files / 5534 tests.

call-graph.ts is now 5,425 → 4,741 lines across six extracted sibling modules
(types, extract, external, complexity, cfg, builtins). No feature, no dependency, no LLM.
