+++ openspec/changes/modularize-call-graph-builder/tasks.md
@@ -4,17 +4,18 @@
 - [x] `analyzer` spec: ADD `StableCallGraphBarrel` requirement (extraction preserves public import surface + behavior)
 
 ## Implementation (taken in safe slices, behind the stable barrel)
-- [x] Capture a byte-level graph snapshot as the before/after regression oracle — a multi-language `CallGraphBuilder.build()` run + the moved pure helpers, serialized canonically and SHA-256'd. Grown per slice to cover each slice's outputs: node `docstring` + `signature` (slice 3), external nodes across `externalKind` http/db/unknown (slice 3b), node `cyclomaticComplexity` > 1 via branchy fixtures (slice 3c), and the full per-function `cfgs` overlay (blocks/edges/defUse/params) incl. arrow + decorated-Python body-digging (slice 3d). Current baseline `34c7bce5…`.
+- [x] Capture a byte-level graph snapshot as the before/after regression oracle — a multi-language `CallGraphBuilder.build()` run + the moved pure helpers, serialized canonically and SHA-256'd. Grown per slice to cover each slice's outputs: node `docstring` + `signature` (slice 3), external nodes across `externalKind` http/db/unknown (slice 3b), node `cyclomaticComplexity` > 1 via branchy fixtures (slice 3c), the full per-function `cfgs` overlay (blocks/edges/defUse/params) incl. arrow + decorated-Python body-digging (slice 3d), and ignored-builtin calls (print/len/JSON/Math) so `isIgnoredCallee`'s drop-or-keep outcome is captured (slice 3e). Current baseline `5fbe0719…`.
 - [x] **Slice 1:** Extract `call-graph-types.ts` (types, edge model, distance helpers, layer helpers). call-graph.ts 5,425 → 5,150 lines.
 - [x] **Slice 3:** Extract `call-graph-extract.ts` (extractDocstringBefore, extractDeclaration) — safest small slice (two pure string-scanning helpers, file-internal, zero deps). call-graph.ts 5,150 → 4,951 lines. NOT re-exported (never on the public surface — imported back only).
 - [x] **Slice 3b:** Extract `call-graph-external.ts` (classifyExternal + EXTERNAL_* tables + getOrCreateExternalNode) — an additional clean section-banner seam (the proposal's module list is illustrative; this is "EXTERNAL NODE HELPER"). Pure, file-internal; only `getOrCreateExternalNode` is imported back, `classifyExternal`/the tables stay private. call-graph.ts 4,951 → 4,887 lines. Removed the now-unused internal `ExternalKind` import (still re-exported on the barrel).
 - [x] **Slice 3c:** Extract `call-graph-complexity.ts` (computeCyclomaticComplexity + CC_PATTERN_* tables) — the "CYCLOMATIC COMPLEXITY" section banner; a pure, dependency-free regex estimator. `computeCyclomaticComplexity` WAS exported (though no external importer), so it is imported back AND re-exported on the barrel to preserve the surface exactly; the patterns stay private. call-graph.ts 4,887 → 4,879 lines.
 - [x] **Slice 3d:** Extract `call-graph-cfg.ts` (buildCfgFor) — the "CFG / DATA-FLOW OVERLAY HELPER" section; a pure fail-soft wrapper around `buildFunctionCfg` that body-resolves declaration wrappers (const-arrow, decorated-Python). File-internal (referenced only in test COMMENTS, not imported), so imported back, not re-exported; the now-unused `buildFunctionCfg` import was dropped from call-graph.ts (`FunctionCfg`/`CfgNode` still used). call-graph.ts 4,879 → 4,846 lines.
+- [x] **Slice 3e:** Extract `call-graph-builtins.ts` (the `*_IGNORED` callee tables + `IGNORED_BY_LANGUAGE`/`ALL_IGNORED_CALLEES` + `isIgnoredCallee` + `SELF_CALL_RECEIVERS` + `isSelfReceiver`) — the cohesive callee-filtering sub-part of the grab-bag CONSTANTS section. Pure data + string predicates, zero deps, zero state. File-internal; only the two predicates are imported back, the tables stay private. `HUB_THRESHOLD` and the style-tally helper stay in place. call-graph.ts 4,846 → 4,741 lines.
 - [ ] **Slice 4:** Extract `call-graph-dispatch.ts` (dedupeOverlappingCalls, synthesizeJavaSuperCalls, safeQuery + dispatch synthesis)
 - [ ] **Slice 5:** Extract `grammar-loader.ts` (grammar cache/load, warnUnavailable, __resetGrammarCacheForTests)
 - [ ] **Slice 6:** Extract `call-graph-nodes.ts` (ensureUniqueNodeIds, materializeCfgByNodeId, findEnclosingFunction, linkCodeToInfra) — the "ATTRIBUTION HELPER" section; deferred behind the smaller slices because findEnclosingFunction is a 27-fan-in hub and materializeCfgByNodeId/linkCodeToInfra touch the CFG side-table + IaC (more coupling)
-- [x] Re-export every moved symbol from `call-graph.ts` so no importer changes — slice 1 re-exports the public type/edge model (`RawEdge`/`CALL_DISTANCE_FALLBACK` kept internal); slices 3/3b/3d moved file-internal helpers (no re-export needed); slice 3c re-exports the previously-public `computeCyclomaticComplexity`; repeated per later slice
-- [x] Verify (slices 1, 3, 3b, 3c, 3d): zero edits to the 155 importers (tsc compiled them all clean); export surface byte-identical (multi-line-aware diff, no add/remove); analyzer + full suite green (279 files / 5534 tests); before/after snapshot byte-identical (slice 1 `131ba4c6…`, slice 3 `58107ac0…`, slice 3b `3a118017…`, slice 3c `7b765f31…`, slice 3d `34c7bce5…`). `stable call-graph barrel` test locks the invariant for later slices.
+- [x] Re-export every moved symbol from `call-graph.ts` so no importer changes — slice 1 re-exports the public type/edge model (`RawEdge`/`CALL_DISTANCE_FALLBACK` kept internal); slices 3/3b/3d/3e moved file-internal helpers (no re-export needed); slice 3c re-exports the previously-public `computeCyclomaticComplexity`; repeated per later slice
+- [x] Verify (slices 1, 3, 3b, 3c, 3d, 3e): zero edits to the 155 importers (tsc compiled them all clean); export surface byte-identical (multi-line-aware diff, no add/remove); analyzer + full suite green (279 files / 5534 tests); before/after snapshot byte-identical (slice 1 `131ba4c6…`, slice 3 `58107ac0…`, slice 3b `3a118017…`, slice 3c `7b765f31…`, slice 3d `34c7bce5…`, slice 3e `5fbe0719…`). `stable call-graph barrel` test locks the invariant for later slices.
 
 ## Verification
 - [x] No new feature, dependency, LLM call, or persisted artifact
